requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
python-dotenv==1.0.1
lightgbm==4.3.0
scikit-learn==1.4.2
pandas==2.2.2
numpy==1.26.4
upstash-redis==1.1.0
kafka-python==2.0.2
pydantic==2.7.1
httpx==0.27.0
sqlalchemy==2.0.30
xgboost==2.1.1
APScheduler==3.10.4
scipy==1.13.1
Jinja2==3.1.4
reportlab==4.2.2
Faker==26.0.0
redis==5.0.7
pytz==2024.1